+++ ReleaseNotes.md
@@ -2,16 +2,25 @@
 
 # Release 0.23.3
 
+## Dependencies
+
+* Serenity BDD 3.2.4
+* RestAssured 5.0.1
+* Selenium 4.1.3
+* Appium 8.0.0
+
 ## Bugfixes
 
 * TGR-536: Filtering on the website no longer splits up message-pairs
 * TGR-534: When a shutdown is triggered during Startup External-Jars are now terminated correctly
+* TGR-486: Requirements are now correctly reported when using the Tiger maven plugin to create the serenity report
 
 ## Features
 
-* TGR-541: TigerProxy WebUI now displays message timestamp in the menu
-* TGR-534: Traffic can now be uploaded from the TigerProxy WebUI
+* TGR-541: Tiger Proxy WebUI now displays message timestamp in the menu
+* TGR-534: Traffic can now be uploaded from the Tiger Proxy WebUI
 * TGR-548: JEXL Debugging/RbelPath Debugging: help text is now collapsible
+* TGR-547: Tiger Workflow UI now links the scenarios in the feature list section in the sidebar with the scenario and its steps shown in the central execution pane for easier navigation
 
 # Release 0.23.2
 
@@ -25,7 +34,7 @@
 ## Breaking changes
 
 * TGR-487: glue code "TGR current response at {string} matches as {word}:" now accepts an enums (JSON|XML) instead of strings and deprecated glue code "TGR current response at {string} matches as {word}" is removed
-* TGR-493: The Test library no longer produces a test jar as we moved all BDD stuff to src/main, thus including the tiger-test-lib jar will suffice. Please adapt your pom.xml to not anymore depend on tiger-test-lib test-jar
+* TGR-493: The Test library no longer produces a test jar as we moved all BDD stuff to src/main, thus including the tiger-test-lib jar will suffice. Please adapt your pom.xml to **not any more depend on tiger-test-lib test-jar**
 
 ## Features